[grape] provide patch() implementation; Monkey module patches Grape
