+++ lib/ddtrace/monkey.rb
@@ -5,14 +5,21 @@ require 'thread'
 # patching code, which is required on demand, when patching.
 require 'ddtrace/contrib/active_record/patcher'
 require 'ddtrace/contrib/elasticsearch/patcher'
-require 'ddtrace/contrib/http/patcher'
+require 'ddtrace/contrib/grape/patcher'
 require 'ddtrace/contrib/redis/patcher'
+require 'ddtrace/contrib/http/patcher'
 
 module Datadog
   # Monkey is used for monkey-patching 3rd party libs.
   module Monkey
     @patched = []
-    @autopatch_modules = { elasticsearch: true, http: true, redis: true, active_record: false }
+    @autopatch_modules = {
+      elasticsearch: true,
+      http: true,
+      redis: true,
+      grape: true,
+      active_record: false
+    }
     # Patchers should expose 2 methods:
     # - patch, which applies our patch if needed. Should be idempotent,
     #   can be call twice but should just do nothing the second time.
@@ -21,6 +28,7 @@ module Datadog
     @patchers = { elasticsearch: Datadog::Contrib::Elasticsearch::Patcher,
                   http: Datadog::Contrib::HTTP::Patcher,
                   redis: Datadog::Contrib::Redis::Patcher,
+                  grape: Datadog::Contrib::Grape::Patcher,
                   active_record: Datadog::Contrib::ActiveRecord::Patcher }
     @mutex = Mutex.new
